separated worker module
